pyproject.toml
[tool.poetry]
name = "amitsegalaisummarizer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.4.1"
python-dotenv = "^1.0.1"
telethon = "^1.34.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
